Add <Indicator /> and <IndicatorGroup /> component
`<IndicatorGroup />` takes a function which you use to pass props to
the generated children. These children should receive an index
prop, `i`. It also takes an `activeIndicator` prop which is
used to match against the children and supply an `active` prop
to them.

`<Indicator />` is the default component used in
`<IndicatorGroup />` if you don't supply another