Add `<SocialLinks />` component
Component which renders social media sharing buttons based on props passed in.
By default, it will render every an icon for the base set of links used at Appear Here. Change the links by passing in a `platforms` array, where each item in the array has a `name` which maps to an Icon and a `shareUrl` which is an ES6 template string