Improve responsiveness of EditorialCard components
Due to server side rendering not being able to take into account the width available to a component on initial render, the user will see the alternate, vertical layout of editorial cards once the page has loaded but before the container query polyfill has had a chance to run. This removes the vertical layout in favour of better showing/hiding of content. This does result in lesss information being shown to the user, but this is a necessary trade off required to not break the design. Content that should be visible on rendered is animated in so as to reduce the jankiness between the initial and final render