Add `<BtnGroup />` component
Enable related actions to sit along side each other. Includes some minor
changes to `<Btn />` which allow parent components to set the props back
to their default state. This allows `<BtnGroup />` to override any
explicitly set props on the `<Btn />` with it's own preferences. This is
preferred to ensure the `<Btn />`s conform to our design system where
buttons inside the group must be of the same context and priority.
`<BtnGroup />` also forces `<Btn />` to be the default variant as a stop
gap solution until we require `hollow` or `subtle` groups