Remove `<Banner />` component in favour of `<Panel />`/`<DismissablePanel />`
Banner is poorly named as it describes one particular context in which
the component has been used. This removes banner in favour of a `<Panel
/>` and `<DismissablePanel />` component, which abstracts out the
generic layout of the banner and the addition of a close button into
separate components. This also adds `<Notification />` as an alias for
`<DismissablePanel />` in order to show intent when using the component