Allow `<TabBar />` active marker offset to be turned off
`<TabBar />` currently enforces that the active marker is always -1px
away from the bottom of the outer most div. This is to ensure the active
marker sits on top of the `<TabBar />`'s default border. In scenario's
where the parent component wishes to supply its own border, the active
marker should sit on top of the parent's border. This change enables
that by making it possible to turn the offset off via props