Fix 3rd party dependency on merge-recursive
The short-id library has not been updated in 7 years and has a dependency on the merge-recursive
library, which GitHub just issued a high-severity security warning for. To fix the security
vulnerability on our end we went with a similar library called shortid (without the dash), which
has been more recently updated and does not have a subdependency on merge-recursive. Its only
dependency is the library nanoid, which itself has no dependencies.