PG-2775 remove map from bloom
As the map is only used in two places in web, but currently counted towards the package it seemed a waste to continue using it in bloom when the map could just be transffered over to the main web repo.

By removing map from bloom, we hope to overall improve the performace on our site on the pages which do not use a map, like the homepage or spaceforideas.